+++ lib/guard/ronn/runner.rb
@@ -6,8 +6,9 @@ module Guard
       class << self
         include Formatter
 
-        def run(paths, options={})
+        def run(paths, options = {})
           return false if paths.empty?
+
           message = options[:message] || "Running: #{paths.join(' ')}"
           UI.info(message, :reset => true)
           success = system(ronn_command(paths, options))
@@ -16,10 +17,10 @@ module Guard
 
       private
 
-        def ronn_command(paths, options={})
+        def ronn_command(paths, options = {})
           cmd_parts = []
-          cmd_parts << "bundle exec" if bundler? && options[:bundler] != false
-          cmd_parts << "ronn"
+          cmd_parts << 'bundle exec' if bundler? && options[:bundler] != false
+          cmd_parts << 'ronn'
           cmd_parts << options[:cli] if options[:cli]
           cmd_parts << paths.join(' ')